fix bugs

+++ aispace/constants.py
@@ -44,8 +44,8 @@ LIST_OF_CLASSLABEL = 'LIST_OF_CLASSLABEL'
 STRING = "STRING"
 
 FEATURE_MAPPING = {
-    LIST_OF_INT: tfds.features.Sequence(tf.float32),
-    INT: tf.float32,
+    LIST_OF_INT: tfds.features.Sequence(tf.int32),
+    INT: tf.int32,
 }
 
 # Task
